Move utils/fix-rules.py to ssg/_utils_fixes.py
