+++ ssg/__init__.py
@@ -22,6 +22,7 @@ import ssg._stig as stig
 import ssg._utils as utils
 import ssg._utils_count as utils_count
 import ssg._utils_duplicates as utils_duplicates
+import ssg._utils_fixes as utils_fixes
 import ssg._utils_overlay as utils_overlay
 import ssg._utils_shadowed as utils_shadowed
 import ssg._xccdf as xccdf
